Colorize blocker rows.

+++ pkg/apis/api/types.go
@@ -389,9 +389,9 @@ type TestFailureAnalysis struct {
 	// FailureCount is the total number of times this test failed in the payloads queried.
 	FailureCount int `json:"failure_count"`
 
-	// BlockerScore represents our confidence this is a blocker, ranges from 0.0 -> 1.0, with 1.0 being near
+	// BlockerScore represents our confidence this is a blocker, ranges from 0 -> 100, with 100 being near
 	// certain this is a payload blocker.
-	BlockerScore float64 `json:"blocker_score"`
+	BlockerScore int `json:"blocker_score"`
 
 	// BlockerScoreReasons explain to humans why the blocker_score was given.
 	BlockerScoreReasons []string `json:"blocker_score_reasons"`
